requirements.txt.txt
pwntools
sys
optparse